💡 refactor(state): PaymentSection에서 cart totals 및 주문 로직 직접 처리하도록 개선
PaymentSection 컴포넌트가 cart 상태와 주문 완료 로직을 내부에서 직접 처리하도록 변경하여 CartSidebar로부터의 props 의존성을 제거함.

변경 내용:
- PaymentSection에서 totals, onCompleteOrder props 제거
- useAtom으로 cartAtom, selectedCouponAtom 직접 사용
- useCart, useCartTotals, useProduct, useNotification 훅을 PaymentSection 내부에서 직접 호출
- CartSidebar → PaymentSection 간 props 전달 제거
- 주문 완료 처리 로직을 PaymentSection 내부로 이동
