+++ src/advanced/components/customer/PaymentSection.tsx
@@ -1,9 +1,40 @@
-interface PaymentSectionProps {
-  totals: { totalBeforeDiscount: number; totalAfterDiscount: number };
-  onCompleteOrder: () => void;
-}
+import { useAtom } from "jotai";
+import { useCallback } from "react";
+import { cartAtom } from "../../atoms/cartAtoms";
+import { selectedCouponAtom } from "../../atoms/couponAtoms";
+import { useCart } from "../../hooks/useCart";
+import { useCartTotals } from "../../hooks/useCartTotals";
+import { useProduct } from "../../hooks/useProduct";
+import { useNotification } from "../../hooks/useNotification";
+import { generateOrderNumber } from "../../utils/orderUtils";
 
-export const PaymentSection = ({ totals, onCompleteOrder }: PaymentSectionProps) => {
+export const PaymentSection = () => {
+  /** 장바구니 상태 - Jotai 사용 */
+  const [cart, setCart] = useAtom(cartAtom);
+
+  /** 선택된 쿠폰 - Jotai 사용 */
+  const [selectedCoupon, setSelectedCoupon] = useAtom(selectedCouponAtom);
+
+  /** 상품 데이터 - useProduct hook 사용 */
+  const { products } = useProduct();
+
+  /** 장바구니 관련 actions - useCart hook 사용 */
+  const { calculateItemTotal } = useCart({ products });
+
+  /** 장바구니 총액 계산 - 직접 hook 사용 */
+  const totals = useCartTotals({ calculateItemTotal });
+
+  /** 알림 표시 */
+  const { showToast } = useNotification();
+
+  /** 주문 완료 */
+  const completeOrder = useCallback(() => {
+    const orderNumber = generateOrderNumber();
+    showToast(`주문이 완료되었습니다. 주문번호: ${orderNumber}`, "success");
+    setCart([]);
+    setSelectedCoupon(null);
+  }, [showToast, setCart, setSelectedCoupon]);
+
   const { totalBeforeDiscount, totalAfterDiscount } = totals;
 
   return (
@@ -30,7 +61,7 @@ export const PaymentSection = ({ totals, onCompleteOrder }: PaymentSectionProps)
       </div>
 
       {/* 결제 버튼 */}
-      <button onClick={onCompleteOrder} className="w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors">
+      <button onClick={completeOrder} className="w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors">
         {totalAfterDiscount.toLocaleString()}원 결제하기
       </button>
